+++ src/com/polymart/model/ChiTietHoaDonThanhToanModel.java
@@ -1,34 +1,34 @@
 package com.polymart.model;
 
 public class ChiTietHoaDonThanhToanModel extends AbstractModel {
-    private Long chiTietSanPham;
-    private Long hoaDonThanhToan;
+    private Integer chiTietSanPham;
+    private Integer hoaDonThanhToan;
     private Integer soLuong;
-    private Integer giamGiaThem;
+    private Long giamGiaThem;
 
     public ChiTietHoaDonThanhToanModel() {
     }
 
-    public ChiTietHoaDonThanhToanModel(Long chiTietSanPham, Long hoaDonThanhToan, Integer soLuong, Integer giamGiaThem) {
+    public ChiTietHoaDonThanhToanModel(Integer chiTietSanPham, Integer hoaDonThanhToan, Integer soLuong, Long giamGiaThem) {
         this.chiTietSanPham = chiTietSanPham;
         this.hoaDonThanhToan = hoaDonThanhToan;
         this.soLuong = soLuong;
         this.giamGiaThem = giamGiaThem;
     }
 
-    public Long getChiTietSanPham() {
+    public Integer getChiTietSanPham() {
         return chiTietSanPham;
     }
 
-    public void setChiTietSanPham(Long chiTietSanPham) {
+    public void setChiTietSanPham(Integer chiTietSanPham) {
         this.chiTietSanPham = chiTietSanPham;
     }
 
-    public Long getHoaDonThanhToan() {
+    public Integer getHoaDonThanhToan() {
         return hoaDonThanhToan;
     }
 
-    public void setHoaDonThanhToan(Long hoaDonThanhToan) {
+    public void setHoaDonThanhToan(Integer hoaDonThanhToan) {
         this.hoaDonThanhToan = hoaDonThanhToan;
     }
 
@@ -40,11 +40,11 @@ public class ChiTietHoaDonThanhToanModel extends AbstractModel {
         this.soLuong = soLuong;
     }
 
-    public Integer getGiamGiaThem() {
+    public Long getGiamGiaThem() {
         return giamGiaThem;
     }
 
-    public void setGiamGiaThem(Integer giamGiaThem) {
+    public void setGiamGiaThem(Long giamGiaThem) {
         this.giamGiaThem = giamGiaThem;
     }
